Update surveys.json (#17)
Removed Tech U - Learning and Development Roles. Tech U has broken this one out into a few different roles, so this is not required any longer for a team member to select. We can keep the json file out there, but we don't want team members to select that survey any longer.